fix: check supertraits for associated type instance check
