+++ main/test/flix/Test.Dec.Assoc.Type.Super.flix
@@ -1,0 +1,12 @@
+mod Test.Dec.Assoc.Type.Super {
+    trait C[a] {
+        type T
+    }
+
+    trait D[a] with C[a] {
+    }
+
+    trait E[a] with D[a] {
+        pub def foo(x: a): C.T[a]
+    }
+}
